Move PseudoElementRequest to its own file https://bugs.webkit.org/show_bug.cgi?id=268048 rdar://121565442
Reviewed by Simon Fraser.

We shouldn't need to import ElementRuleCollector.h to use this class.

* Source/WebCore/WebCore.xcodeproj/project.pbxproj:
* Source/WebCore/style/ElementRuleCollector.h:
(WebCore::Style::PseudoElementRequest::PseudoElementRequest): Deleted.
* Source/WebCore/style/PseudoElementRequest.h: Added.
(WebCore::Style::PseudoElementRequest::PseudoElementRequest):

Canonical link: https://commits.webkit.org/273470@main
